cleaned up force
